refactor: delete old reports charts + retire LineChart tests, migrate Pie/Bar tests to Atlas components
- Deleted src/components/reports/charts/{PieChart,BarChart,LineChart}.vue
- Deleted tests for LineChart (incompatible data contract; covered by chart-config.spec.ts)
- Migrated tests/component and tests/unit Pie/Bar specs to AtlasPieChart/AtlasBarChart:
  import path, vi.mock path, CSS selectors (.pie-chart-container→.atlas-pie-chart,
  .bar-chart-container→.atlas-bar-chart)
- NOTE: src/utils/chart-config.ts shim retained — 4 datasources/charts components still
  import from @/utils/chart-config and must be migrated before the shim can be removed
