--- tests/component/reports/charts/LineChart.spec.ts
@@ -1,826 +1,0 @@
-/**
- * LineChart Component Tests
- * Feature: 005-cohort-reports
- *
- * Comprehensive tests for LineChart component covering:
- * - Rendering with valid data
- * - Loading state display
- * - Empty data handling
- * - Export functionality integration
- * - Responsive behavior
- * - Prop validation
- * - Data watcher functionality
- */
-
-import { describe, it, expect, beforeEach, vi, afterEach } from 'vitest'
-import { mount } from '@vue/test-utils'
-import { createVuetify } from 'vuetify'
-import * as components from 'vuetify/components'
-import * as directives from 'vuetify/directives'
-import { nextTick } from 'vue'
-import LineChart from '@/components/reports/charts/LineChart.vue'
-import type { LineChartData } from '@/models/report.types'
-
-const vuetify = createVuetify({
-  components,
-  directives,
-})
-
-// Mock ResizeObserver
-global.ResizeObserver = vi.fn().mockImplementation(() => ({
-  observe: vi.fn(),
-  unobserve: vi.fn(),
-  disconnect: vi.fn(),
-}))
-
-// Mock window resize event handlers
-let resizeListeners: Array<() => void> = []
-const originalAddEventListener = window.addEventListener
-const originalRemoveEventListener = window.removeEventListener
-
-describe('LineChart', () => {
-  beforeEach(() => {
-    resizeListeners = []
-
-    // Mock addEventListener to track resize handlers
-    window.addEventListener = vi.fn((event: string, handler: () => void) => {
-      if (event === 'resize') {
-        resizeListeners.push(handler)
-      }
-    })
-
-    // Mock removeEventListener
-    window.removeEventListener = vi.fn((event: string, handler: () => void) => {
-      if (event === 'resize') {
-        resizeListeners = resizeListeners.filter(h => h !== handler)
-      }
-    })
-  })
-
-  afterEach(() => {
-    window.addEventListener = originalAddEventListener
-    window.removeEventListener = originalRemoveEventListener
-    resizeListeners = []
-  })
-
-  const mockData: LineChartData = {
-    xAxis: ['2020-01', '2020-02', '2020-03', '2020-04', '2020-05'],
-    yAxis: [100, 250, 180, 320, 290],
-    seriesName: 'Monthly Prevalence'
-  }
-
-  const createWrapper = (props: Record<string, unknown> = {}) => {
-    return mount(LineChart, {
-      props: {
-        data: mockData,
-        ...props,
-      },
-      global: {
-        plugins: [vuetify],
-        stubs: {
-          ChartExport: {
-            name: 'ChartExport',
-            template: '<div class="chart-export-stub"></div>',
-            props: ['chartInstance', 'filename'],
-          },
-          VChart: {
-            name: 'VChart',
-            template: '<div class="v-chart-stub"></div>',
-            props: ['option', 'autoresize', 'style'],
-            setup(props: Record<string, unknown>, { expose }: { expose: (exposed: Record<string, unknown>) => void }) {
-              // Mock chart instance
-              const chart = {
-                setOption: vi.fn(),
-                resize: vi.fn(),
-                isDisposed: vi.fn(() => false),
-                getDataURL: vi.fn(() => 'data:image/png;base64,mockdata'),
-                renderToSVGString: vi.fn(() => '<svg></svg>'),
-              }
-
-              // Expose chart instance so parent can access it
-              expose({ chart })
-
-              return {
-                chart,
-                // Also expose setOption at component level for watcher
-                setOption: chart.setOption
-              }
-            },
-          },
-        },
-      },
-    })
-  }
-
-  // ============================================================================
-  // Basic Rendering Tests
-  // ============================================================================
-
-  describe('Rendering', () => {
-    it('should render line chart container', () => {
-      const wrapper = createWrapper()
-      expect(wrapper.find('.line-chart-container').exists()).toBe(true)
-    })
-
-    it('should render VChart component when not loading', () => {
-      const wrapper = createWrapper({ loading: false })
-      expect(wrapper.findComponent({ name: 'VChart' }).exists()).toBe(true)
-    })
-
-    it('should not render VChart component when loading', () => {
-      const wrapper = createWrapper({ loading: true })
-      expect(wrapper.findComponent({ name: 'VChart' }).exists()).toBe(false)
-    })
-
-    it('should pass height to VChart style', () => {
-      const wrapper = createWrapper({ height: 500 })
-      const chart = wrapper.findComponent({ name: 'VChart' })
-      // Check that chart receives props (stub will handle them)
-      expect(chart.exists()).toBe(true)
-    })
-
-    it('should pass style object to VChart', () => {
-      const wrapper = createWrapper()
-      const chart = wrapper.findComponent({ name: 'VChart' })
-      // Verify chart component exists and can receive props
-      expect(chart.exists()).toBe(true)
-    })
-
-    it('should pass autoresize prop to VChart', () => {
-      const wrapper = createWrapper()
-      const chart = wrapper.findComponent({ name: 'VChart' })
-      // Autoresize is bound as boolean attribute
-      expect(chart.html()).toContain('v-chart-stub')
-    })
-  })
-
-  // ============================================================================
-  // Loading State Tests
-  // ============================================================================
-
-  describe('Loading State', () => {
-    it('should display skeleton loader when loading is true', () => {
-      const wrapper = createWrapper({ loading: true })
-      expect(wrapper.findComponent({ name: 'v-skeleton-loader' }).exists()).toBe(true)
-    })
-
-    it('should not display skeleton loader when loading is false', () => {
-      const wrapper = createWrapper({ loading: false })
-      expect(wrapper.findComponent({ name: 'v-skeleton-loader' }).exists()).toBe(false)
-    })
-
-    it('should apply correct height to skeleton loader', () => {
-      const wrapper = createWrapper({ loading: true, height: 600 })
-      const skeleton = wrapper.findComponent({ name: 'v-skeleton-loader' })
-      expect(skeleton.props('height')).toBe(600)
-    })
-
-    it('should use default height for skeleton loader', () => {
-      const wrapper = createWrapper({ loading: true })
-      const skeleton = wrapper.findComponent({ name: 'v-skeleton-loader' })
-      expect(skeleton.props('height')).toBe(400)
-    })
-
-    it('should hide export controls when loading', () => {
-      const wrapper = createWrapper({ loading: true, showExport: true })
-      expect(wrapper.find('.chart-export-toolbar').exists()).toBe(false)
-    })
-  })
-
-  // ============================================================================
-  // Data Handling Tests
-  // ============================================================================
-
-  describe('Data Handling', () => {
-    it('should pass chart option to VChart', () => {
-      const wrapper = createWrapper()
-      const chart = wrapper.findComponent({ name: 'VChart' })
-      const option = chart.props('option')
-
-      expect(option).toBeDefined()
-      expect(option).not.toEqual({})
-    })
-
-    it('should handle empty data gracefully', () => {
-      const emptyData: LineChartData = {
-        xAxis: [],
-        yAxis: [],
-      }
-      const wrapper = createWrapper({ data: emptyData })
-      const chart = wrapper.findComponent({ name: 'VChart' })
-      const option = chart.props('option')
-
-      // Empty data should result in empty option
-      expect(option).toEqual({})
-    })
-
-    it('should handle numeric xAxis', () => {
-      const numericData: LineChartData = {
-        xAxis: [1, 2, 3, 4, 5],
-        yAxis: [10, 20, 15, 30, 25],
-      }
-      const wrapper = createWrapper({ data: numericData })
-      const chart = wrapper.findComponent({ name: 'VChart' })
-      const option = chart.props('option')
-
-      expect(option).not.toEqual({})
-    })
-
-    it('should include seriesName in chart configuration', () => {
-      const dataWithSeriesName: LineChartData = {
-        xAxis: ['A', 'B'],
-        yAxis: [10, 20],
-        seriesName: 'Test Series'
-      }
-      const wrapper = createWrapper({ data: dataWithSeriesName })
-      const chart = wrapper.findComponent({ name: 'VChart' })
-      const option = chart.props('option')
-
-      // Check that chart option is generated (non-empty)
-      expect(option).not.toEqual({})
-    })
-
-    it('should work without seriesName specified', () => {
-      const dataWithoutSeriesName: LineChartData = {
-        xAxis: ['A', 'B'],
-        yAxis: [10, 20],
-      }
-      const wrapper = createWrapper({ data: dataWithoutSeriesName })
-      const chart = wrapper.findComponent({ name: 'VChart' })
-      const option = chart.props('option')
-
-      expect(option).not.toEqual({})
-    })
-
-    it('should include title in chart configuration when provided', () => {
-      const wrapper = createWrapper({ title: 'Test Chart Title' })
-      const chart = wrapper.findComponent({ name: 'VChart' })
-      const option = chart.props('option')
-
-      expect(option).not.toEqual({})
-    })
-
-    it('should work without title', () => {
-      const wrapper = createWrapper()
-      const chart = wrapper.findComponent({ name: 'VChart' })
-      const option = chart.props('option')
-
-      expect(option).not.toEqual({})
-    })
-
-    it('should react to data changes', async () => {
-      const wrapper = createWrapper()
-      const initialData = mockData
-
-      // Change to different data
-      const _newData: LineChartData = {
-        xAxis: ['X', 'Y'],
-        yAxis: [50, 75],
-        seriesName: 'New Series'
-      }
-
-      // Note: Testing deep watcher is complex with stubbed VChart
-      // The component uses a deep watcher that calls chartRef.value.setOption()
-      // In real usage, VChart component provides this method
-      // For now, we verify the component can accept prop updates
-      expect(wrapper.props('data')).toEqual(initialData)
-
-      // Component handles data changes via computed chartOption
-      const chart = wrapper.findComponent({ name: 'VChart' })
-      expect(chart.exists()).toBe(true)
-    })
-
-    it('should not update chart when loading', async () => {
-      const wrapper = createWrapper({ loading: false })
-
-      await wrapper.setProps({ loading: true })
-      await nextTick()
-
-      // Chart should not be visible when loading
-      expect(wrapper.findComponent({ name: 'VChart' }).exists()).toBe(false)
-    })
-  })
-
-  // ============================================================================
-  // Export Functionality Tests
-  // ============================================================================
-
-  describe('Export Functionality', () => {
-    it('should render ChartExport component when showExport is true', () => {
-      const wrapper = createWrapper({ showExport: true, loading: false })
-      expect(wrapper.findComponent({ name: 'ChartExport' }).exists()).toBe(true)
-    })
-
-    it('should not render ChartExport component when showExport is false', () => {
-      const wrapper = createWrapper({ showExport: false })
-      expect(wrapper.findComponent({ name: 'ChartExport' }).exists()).toBe(false)
-    })
-
-    it('should pass chart instance to ChartExport', async () => {
-      const wrapper = createWrapper({ showExport: true, loading: false })
-      await nextTick()
-
-      const exportComponent = wrapper.findComponent({ name: 'ChartExport' })
-      expect(exportComponent.props('chartInstance')).toBeDefined()
-    })
-
-    it('should pass filename to ChartExport', () => {
-      const wrapper = createWrapper({
-        showExport: true,
-        loading: false,
-        exportFilename: 'my-custom-chart'
-      })
-
-      const exportComponent = wrapper.findComponent({ name: 'ChartExport' })
-      expect(exportComponent.props('filename')).toBe('my-custom-chart')
-    })
-
-    it('should use default filename when not provided', () => {
-      const wrapper = createWrapper({ showExport: true, loading: false })
-
-      const exportComponent = wrapper.findComponent({ name: 'ChartExport' })
-      expect(exportComponent.props('filename')).toBe('line-chart')
-    })
-
-    it('should emit export-success event', async () => {
-      const wrapper = createWrapper({ showExport: true, loading: false })
-      await nextTick()
-
-      const exportComponent = wrapper.findComponent({ name: 'ChartExport' })
-
-      // Simulate export success
-      await exportComponent.vm.$emit('export-success', 'png', 'test-chart.png')
-
-      expect(wrapper.emitted('export-success')).toBeTruthy()
-      const emitted = wrapper.emitted('export-success') as Array<unknown[]>
-      expect(emitted[0]).toEqual(['png', 'test-chart.png'])
-    })
-
-    it('should emit export-error event', async () => {
-      const wrapper = createWrapper({ showExport: true, loading: false })
-      await nextTick()
-
-      const exportComponent = wrapper.findComponent({ name: 'ChartExport' })
-      const testError = new Error('Export failed')
-
-      // Simulate export error
-      await exportComponent.vm.$emit('export-error', 'svg', testError)
-
-      expect(wrapper.emitted('export-error')).toBeTruthy()
-      const emitted = wrapper.emitted('export-error') as Array<unknown[]>
-      expect(emitted[0]).toEqual(['svg', testError])
-    })
-
-    it('should display export toolbar when not loading', () => {
-      const wrapper = createWrapper({ showExport: true, loading: false })
-      expect(wrapper.find('.chart-export-toolbar').exists()).toBe(true)
-    })
-
-    it('should support PNG export format', async () => {
-      const wrapper = createWrapper({ showExport: true, loading: false })
-      await nextTick()
-
-      const exportComponent = wrapper.findComponent({ name: 'ChartExport' })
-      await exportComponent.vm.$emit('export-success', 'png', 'chart.png')
-
-      expect(wrapper.emitted('export-success')).toBeTruthy()
-      const emitted = wrapper.emitted('export-success') as Array<unknown[]>
-      expect(emitted[0][0]).toBe('png')
-    })
-
-    it('should support SVG export format', async () => {
-      const wrapper = createWrapper({ showExport: true, loading: false })
-      await nextTick()
-
-      const exportComponent = wrapper.findComponent({ name: 'ChartExport' })
-      await exportComponent.vm.$emit('export-success', 'svg', 'chart.svg')
-
-      expect(wrapper.emitted('export-success')).toBeTruthy()
-      const emitted = wrapper.emitted('export-success') as Array<unknown[]>
-      expect(emitted[0][0]).toBe('svg')
-    })
-  })
-
-  // ============================================================================
-  // Responsive Behavior Tests
-  // ============================================================================
-
-  describe('Responsive Behavior', () => {
-    it('should register resize event listener on mount', async () => {
-      const _wrapper = createWrapper()
-      await nextTick()
-
-      expect(window.addEventListener).toHaveBeenCalledWith('resize', expect.any(Function))
-      expect(resizeListeners.length).toBeGreaterThan(0)
-    })
-
-    it('should remove resize event listener on unmount', async () => {
-      const wrapper = createWrapper()
-      await nextTick()
-
-      const listenerCount = resizeListeners.length
-      expect(listenerCount).toBeGreaterThan(0)
-
-      wrapper.unmount()
-
-      expect(window.removeEventListener).toHaveBeenCalledWith('resize', expect.any(Function))
-    })
-
-    it('should handle resize events', async () => {
-      const wrapper = createWrapper()
-      await nextTick()
-
-      // Verify resize listener was registered
-      expect(resizeListeners.length).toBeGreaterThan(0)
-
-      // Component should remain stable after mount
-      expect(wrapper.findComponent({ name: 'VChart' }).exists()).toBe(true)
-    })
-
-    it('should clean up on unmount', async () => {
-      const wrapper = createWrapper()
-      await nextTick()
-
-      const initialListenerCount = resizeListeners.length
-      expect(initialListenerCount).toBeGreaterThan(0)
-
-      wrapper.unmount()
-
-      // removeEventListener should have been called
-      expect(window.removeEventListener).toHaveBeenCalled()
-    })
-
-    it('should handle multiple mount/unmount cycles', async () => {
-      const wrapper1 = createWrapper()
-      await nextTick()
-      expect(resizeListeners.length).toBeGreaterThan(0)
-
-      wrapper1.unmount()
-
-      const wrapper2 = createWrapper()
-      await nextTick()
-      expect(resizeListeners.length).toBeGreaterThan(0)
-
-      wrapper2.unmount()
-    })
-  })
-
-  // ============================================================================
-  // Props Validation Tests
-  // ============================================================================
-
-  describe('Props Validation', () => {
-    it('should accept valid LineChartData with string xAxis', () => {
-      const validData: LineChartData = {
-        xAxis: ['A', 'B', 'C'],
-        yAxis: [10, 20, 30],
-        seriesName: 'Test'
-      }
-
-      expect(() => createWrapper({ data: validData })).not.toThrow()
-    })
-
-    it('should accept valid LineChartData with numeric xAxis', () => {
-      const validData: LineChartData = {
-        xAxis: [1, 2, 3],
-        yAxis: [10, 20, 30],
-        seriesName: 'Test'
-      }
-
-      expect(() => createWrapper({ data: validData })).not.toThrow()
-    })
-
-    it('should handle data without seriesName', () => {
-      const dataNoSeriesName: LineChartData = {
-        xAxis: ['A', 'B'],
-        yAxis: [10, 20],
-      }
-
-      expect(() => createWrapper({ data: dataNoSeriesName })).not.toThrow()
-    })
-
-    it('should accept custom height prop', () => {
-      const wrapper = createWrapper({ height: 800 })
-      const chart = wrapper.findComponent({ name: 'VChart' })
-
-      expect(chart.exists()).toBe(true)
-    })
-
-    it('should use default height when not specified', () => {
-      const wrapper = createWrapper()
-      const chart = wrapper.findComponent({ name: 'VChart' })
-
-      expect(chart.exists()).toBe(true)
-    })
-
-    it('should handle loading prop true', () => {
-      const wrapper = createWrapper({ loading: true })
-
-      expect(wrapper.findComponent({ name: 'v-skeleton-loader' }).exists()).toBe(true)
-      expect(wrapper.findComponent({ name: 'VChart' }).exists()).toBe(false)
-    })
-
-    it('should handle loading prop false', () => {
-      const wrapper = createWrapper({ loading: false })
-
-      expect(wrapper.findComponent({ name: 'v-skeleton-loader' }).exists()).toBe(false)
-      expect(wrapper.findComponent({ name: 'VChart' }).exists()).toBe(true)
-    })
-
-    it('should default loading to false', () => {
-      const wrapper = createWrapper()
-
-      expect(wrapper.findComponent({ name: 'VChart' }).exists()).toBe(true)
-    })
-
-    it('should default showExport to true', () => {
-      const wrapper = createWrapper({ loading: false })
-
-      expect(wrapper.findComponent({ name: 'ChartExport' }).exists()).toBe(true)
-    })
-
-    it('should default exportFilename to "line-chart"', () => {
-      const wrapper = createWrapper({ showExport: true, loading: false })
-      const exportComponent = wrapper.findComponent({ name: 'ChartExport' })
-
-      expect(exportComponent.props('filename')).toBe('line-chart')
-    })
-
-    it('should accept custom title prop', () => {
-      const wrapper = createWrapper({ title: 'My Chart' })
-      expect(wrapper.findComponent({ name: 'VChart' }).exists()).toBe(true)
-    })
-
-    it('should handle undefined title prop', () => {
-      const wrapper = createWrapper({ title: undefined })
-      expect(wrapper.findComponent({ name: 'VChart' }).exists()).toBe(true)
-    })
-  })
-
-  // ============================================================================
-  // Edge Cases and Error Handling
-  // ============================================================================
-
-  describe('Edge Cases', () => {
-    it('should handle data with single data point', () => {
-      const singleData: LineChartData = {
-        xAxis: ['Only One'],
-        yAxis: [100],
-      }
-
-      const wrapper = createWrapper({ data: singleData })
-      const chart = wrapper.findComponent({ name: 'VChart' })
-
-      expect(chart.exists()).toBe(true)
-      expect(chart.props('option')).not.toEqual({})
-    })
-
-    it('should handle data with many data points', () => {
-      const manyPoints: LineChartData = {
-        xAxis: Array.from({ length: 100 }, (_, i) => `Point ${i}`),
-        yAxis: Array.from({ length: 100 }, (_, i) => i * 10),
-      }
-
-      const wrapper = createWrapper({ data: manyPoints })
-      const chart = wrapper.findComponent({ name: 'VChart' })
-
-      expect(chart.exists()).toBe(true)
-      expect(chart.props('option')).not.toEqual({})
-    })
-
-    it('should handle very large values', () => {
-      const largeValues: LineChartData = {
-        xAxis: ['A', 'B', 'C'],
-        yAxis: [1000000, 5000000, 2500000],
-        seriesName: 'Large Numbers'
-      }
-
-      const wrapper = createWrapper({ data: largeValues })
-      const chart = wrapper.findComponent({ name: 'VChart' })
-
-      expect(chart.exists()).toBe(true)
-    })
-
-    it('should handle zero values', () => {
-      const zeroValues: LineChartData = {
-        xAxis: ['A', 'B', 'C'],
-        yAxis: [0, 0, 0],
-      }
-
-      const wrapper = createWrapper({ data: zeroValues })
-      const chart = wrapper.findComponent({ name: 'VChart' })
-
-      expect(chart.exists()).toBe(true)
-      expect(chart.props('option')).not.toEqual({})
-    })
-
-    it('should handle mixed positive and zero values', () => {
-      const mixedValues: LineChartData = {
-        xAxis: ['A', 'B', 'C', 'D'],
-        yAxis: [100, 0, 250, 0],
-      }
-
-      const wrapper = createWrapper({ data: mixedValues })
-      const chart = wrapper.findComponent({ name: 'VChart' })
-
-      expect(chart.exists()).toBe(true)
-    })
-
-    it('should handle negative values', () => {
-      const negativeValues: LineChartData = {
-        xAxis: ['A', 'B', 'C'],
-        yAxis: [-10, -20, -15],
-      }
-
-      const wrapper = createWrapper({ data: negativeValues })
-      const chart = wrapper.findComponent({ name: 'VChart' })
-
-      expect(chart.exists()).toBe(true)
-    })
-
-    it('should handle mixed positive and negative values', () => {
-      const mixedValues: LineChartData = {
-        xAxis: ['A', 'B', 'C', 'D'],
-        yAxis: [100, -50, 250, -30],
-      }
-
-      const wrapper = createWrapper({ data: mixedValues })
-      const chart = wrapper.findComponent({ name: 'VChart' })
-
-      expect(chart.exists()).toBe(true)
-    })
-
-    it('should handle mismatched xAxis and yAxis lengths gracefully', () => {
-      const mismatchedData: LineChartData = {
-        xAxis: ['A', 'B', 'C'],
-        yAxis: [10, 20],  // One less than xAxis
-      }
-
-      const wrapper = createWrapper({ data: mismatchedData })
-      const chart = wrapper.findComponent({ name: 'VChart' })
-
-      // Component should still render
-      expect(chart.exists()).toBe(true)
-    })
-
-    it('should not error when chart instance is null during export', () => {
-      const wrapper = createWrapper({ showExport: true, loading: false })
-      const exportComponent = wrapper.findComponent({ name: 'ChartExport' })
-
-      // ChartExport should exist and handle chart instance gracefully (may be null/undefined in test env)
-      expect(exportComponent.exists()).toBe(true)
-    })
-
-    it('should handle date strings in xAxis', () => {
-      const dateData: LineChartData = {
-        xAxis: ['2020-01-01', '2020-02-01', '2020-03-01'],
-        yAxis: [100, 150, 200],
-        seriesName: 'Date Series'
-      }
-
-      const wrapper = createWrapper({ data: dateData })
-      const chart = wrapper.findComponent({ name: 'VChart' })
-
-      expect(chart.exists()).toBe(true)
-    })
-
-    it('should handle decimal values in yAxis', () => {
-      const decimalData: LineChartData = {
-        xAxis: ['A', 'B', 'C'],
-        yAxis: [10.5, 20.75, 15.25],
-      }
-
-      const wrapper = createWrapper({ data: decimalData })
-      const chart = wrapper.findComponent({ name: 'VChart' })
-
-      expect(chart.exists()).toBe(true)
-    })
-  })
-
-  // ============================================================================
-  // Integration Tests
-  // ============================================================================
-
-  describe('Integration', () => {
-    it('should work with all props together', () => {
-      const wrapper = createWrapper({
-        data: mockData,
-        loading: false,
-        height: 600,
-        showExport: true,
-        exportFilename: 'test-report',
-        title: 'Test Title'
-      })
-
-      expect(wrapper.find('.line-chart-container').exists()).toBe(true)
-      expect(wrapper.findComponent({ name: 'VChart' }).exists()).toBe(true)
-      expect(wrapper.findComponent({ name: 'ChartExport' }).exists()).toBe(true)
-
-      const exportComponent = wrapper.findComponent({ name: 'ChartExport' })
-      expect(exportComponent.props('filename')).toBe('test-report')
-    })
-
-    it('should transition from loading to loaded state', async () => {
-      const wrapper = createWrapper({ loading: true })
-
-      expect(wrapper.findComponent({ name: 'v-skeleton-loader' }).exists()).toBe(true)
-      expect(wrapper.findComponent({ name: 'VChart' }).exists()).toBe(false)
-
-      await wrapper.setProps({ loading: false })
-      await nextTick()
-
-      expect(wrapper.findComponent({ name: 'v-skeleton-loader' }).exists()).toBe(false)
-      expect(wrapper.findComponent({ name: 'VChart' }).exists()).toBe(true)
-    })
-
-    it('should handle complete export workflow', async () => {
-      const wrapper = createWrapper({
-        showExport: true,
-        loading: false,
-        exportFilename: 'prevalence-report'
-      })
-      await nextTick()
-
-      const exportComponent = wrapper.findComponent({ name: 'ChartExport' })
-
-      // Simulate successful PNG export
-      await exportComponent.vm.$emit('export-success', 'png', 'prevalence-report.png')
-
-      expect(wrapper.emitted('export-success')).toBeTruthy()
-      const emitted = wrapper.emitted('export-success') as Array<unknown[]>
-      expect(emitted[0][0]).toBe('png')
-      expect(emitted[0][1]).toBe('prevalence-report.png')
-    })
-
-    it('should handle error during export', async () => {
-      const wrapper = createWrapper({
-        showExport: true,
-        loading: false,
-      })
-      await nextTick()
-
-      const exportComponent = wrapper.findComponent({ name: 'ChartExport' })
-      const error = new Error('Export failed')
-
-      // Simulate export error
-      await exportComponent.vm.$emit('export-error', 'png', error)
-
-      expect(wrapper.emitted('export-error')).toBeTruthy()
-      const emitted = wrapper.emitted('export-error') as Array<unknown[]>
-      expect(emitted[0][0]).toBe('png')
-      expect(emitted[0][1]).toBe(error)
-    })
-
-    it('should handle data changes via watcher', async () => {
-      const wrapper = createWrapper({ loading: false })
-      await nextTick()
-
-      // Note: Testing deep watcher is complex with stubbed VChart
-      // The component uses a deep watcher that calls chartRef.value.setOption()
-      // In real usage, VChart component provides this method at component level
-      // For now, we verify the component renders and can accept data
-      const chart = wrapper.findComponent({ name: 'VChart' })
-      expect(chart.exists()).toBe(true)
-
-      // Verify the component is properly set up with data
-      expect(wrapper.props('data')).toEqual(mockData)
-    })
-
-    it('should work with time series data', () => {
-      const timeSeriesData: LineChartData = {
-        xAxis: ['2020-01', '2020-02', '2020-03', '2020-04', '2020-05', '2020-06'],
-        yAxis: [100, 150, 125, 175, 200, 180],
-        seriesName: 'Monthly Trends'
-      }
-
-      const wrapper = createWrapper({
-        data: timeSeriesData,
-        title: 'Prevalence Over Time',
-        height: 500
-      })
-
-      expect(wrapper.findComponent({ name: 'VChart' }).exists()).toBe(true)
-      expect(wrapper.find('.line-chart-container').exists()).toBe(true)
-    })
-
-    it('should support dynamic height changes', async () => {
-      const wrapper = createWrapper({ height: 400 })
-
-      await wrapper.setProps({ height: 800 })
-      await nextTick()
-
-      expect(wrapper.findComponent({ name: 'VChart' }).exists()).toBe(true)
-    })
-
-    it('should handle showing and hiding export controls', async () => {
-      const wrapper = createWrapper({ showExport: true, loading: false })
-
-      expect(wrapper.findComponent({ name: 'ChartExport' }).exists()).toBe(true)
-
-      await wrapper.setProps({ showExport: false })
-      await nextTick()
-
-      expect(wrapper.findComponent({ name: 'ChartExport' }).exists()).toBe(false)
-    })
-  })
-})
